+++ app/sdc-landing/src/hook/Query/Supabase/index.ts
@@ -3,3 +3,4 @@ export * as Discord from './Discord';
 export { default as useGetSession } from './useGetSession';
 export { default as useQueryProfiles } from './useQueryProfiles';
 export { default as useQueryMyProfile } from './useQueryMyProfile';
+export { default as useQueryProfileByUsername } from './useQueryProfileByUsername';
